workaround problems with sqlx named args
This commit adds a workaround for problems with the sqlx library's named
args. sqlx uses a hardcoded list of driver names to determine how to
represent parameters in prepared queries. For example, postgres uses $1,
mysql uses ?, sqlserver uses @, and so on. But since copyist defines a
custom driver name, sqlx falls back to the default ?, which won't work
with some databases. These issues describe the "custom driver" problem:

  https://github.com/jmoiron/sqlx/issues/400
  https://github.com/jmoiron/sqlx/issues/559

Workaround this problem by rebinding the query if the bind type of the
inner driver is different than the default ? character.