enable copyist.Register to be called in init() function
Previously, the recommendation was to call copyist.Register in the TestMain
function, after parsing flags. However, it's more intuitive if it can be
called in an init() function instead.

This commit updates copyist to lazily get the underlying driver, the first
time a copyist connection is opened. Flags have already been parsed at this
point, and a dataSourceName is available, which cleans up some awkward code
to get the sql.Driver that used to be in copyist.Register. Also, Register
no longer needs to return an error, which simplifies usage.