generate custom text file rather than Go code file
Change copyist to generate a custom text file that sits alongside the
test file which invokes copyist. This works better than generated Go code,
since it does not need to be recompiled every time a test is changed. It's
also simpler.

See the header comment for recordingFile for more details on the format of
the file.

Also clarify in the README that copyist enables parallel execution of
test packages during playback.