improve findTestFile heuristics
`findTestFile` was previously returning the first file in the call stack
that ended in `_test.go`, but that heuristic doesn't work nicely when
there is a `_test.go` file that contains utility functions, such as a
`testServer` struct. The new heuristics is to walk up the call stack
looking for the last file ending in `_test.go`.